requirements.txt
torch>=1.6.0
torchvision>=0.7.0
dominate>=2.4.0
visdom>=0.1.8.8
packaging
GPUtil>=1.4.0
scipy
Pillow>=6.1.0
numpy>=1.16.4
kornia
torch==2.7.0+cu118 torchvision==0.22.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
